% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing.R
\name{attachEndpoints}
\alias{attachEndpoints}
\title{attachEndpoints}
\usage{
attachEndpoints(nodes, links, latB, latT, lonL, lonR, numEndpoints,
  bandwidth, bandwidthUnits, idPrefix)
}
\arguments{
\item{nodes}{data.frame(id, type, lon, lat)}

\item{latB}{region bottom latitude limit}

\item{latT}{region top latitude limit}

\item{lonL}{region left longitude limit}

\item{lonR}{region right longitude limit}

\item{numEndpoints}{number of endpoints to be generated}

\item{bandwidth}{bandwidth resources available for each user}

\item{bandwidthUnits}{string for bandwidth units}

\item{idPrefix}{string to be prefixed before the fog node id}

\item{link}{data.frame(from, to, bandwidth, bandwidthUnits, distance,
distanceUnits)}
}
\value{
list(nodes = data.frame, links = data.frame)
}
\description{
Generates endpoints within the selected region, asking for
certain bandwidth resources. And attaches them to the nearest cells.
}
