% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing.R
\name{attachFogEndpoints}
\alias{attachFogEndpoints}
\title{attachFogEndpoints}
\usage{
attachFogEndpoints(nodes, links, latB, latT, lonL, lonR, numNodes,
  properties, bandwidth, bandwidthUnits, idPrefix, dis = -1)
}
\arguments{
\item{nodes}{data.frame(id, type, lon, lat)}

\item{latB}{region bottom latitude limit}

\item{latT}{region top latitude limit}

\item{lonL}{region left longitude limit}

\item{lonR}{region right longitude limit}

\item{numNodes}{number of fog compute nodes to generate}

\item{properties}{list with compute node resource properties}

\item{bandwidth}{bandwidth for link between each server and the switch}

\item{bandwidthUnits}{string for bandwidth units}

\item{idPrefix}{string to be prefixed before the fog node id}

\item{dis}{minimum distance in meters to attach to a cell, if not specified
it attachs the fog nodes to their nearest cells}

\item{link}{data.frame(from, to, bandwidth, bandwidthUnits, distance,
distanceUnits)}

\item{distanceUnits}{string for distance units}
}
\value{
list(nodes = data.frame, links = data.frame)
}
\description{
Creates fog nodes with associated endpoints. The fog nodes are
connected to cells, while the endpoints are attached to the fog nodes.
}
