% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing.R
\name{attachFogNodes}
\alias{attachFogNodes}
\title{attachFogNodes}
\usage{
attachFogNodes(nodes, links, latB, latT, lonL, lonR, numNodes, properties,
  bandwidth, bandwidthUnits, idPrefix)
}
\arguments{
\item{nodes}{data.frame(id, type, lon, lat)}

\item{latB}{region bottom latitude limit}

\item{latT}{region top latitude limit}

\item{lonL}{region left longitude limit}

\item{lonR}{region right longitude limit}

\item{numNodes}{number of fog compute nodes to generate}

\item{properties}{list with compute node resource properties}

\item{bandwidth}{bandwidth for link between each server and the switch}

\item{bandwidthUnits}{string for bandwidth units}

\item{idPrefix}{string to be prefixed before the fog node id}

\item{link}{data.frame(from, to, bandwidth, bandwidthUnits, distance,
distanceUnits)}

\item{distanceUnits}{string for distance units}
}
\value{
list(nodes = data.frame, links = data.frame)
}
\description{
Generates fog computing nodes within a certain region, and
attaches them to the existing infrastructure
}
