% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing.R
\name{attachServers}
\alias{attachServers}
\title{attachServers}
\usage{
attachServers(nodes, links, numServers, bandwidth, bandwidthUnits,
  distance, distanceUnits, switchType, properties, idPrefix)
}
\arguments{
\item{nodes}{data.frame(id, type, lon, lat)}

\item{numServers}{number of servers to be attached}

\item{bandwidth}{bandwidth for link between each server and the switch}

\item{bandwidthUnits}{string for bandwidth units}

\item{distance}{distance beween servers and switches}

\item{distanceUnits}{string for distance units}

\item{switchType}{string to determine the type of switch to attach}

\item{properties}{list with each server properties, e.g., list(cpu=,disk=)}

\item{idPrefix}{string to be prefixed before the server id}

\item{link}{data.frame(from, to, bandwidth, bandwidthUnits, distance,
distanceUnits)}
}
\value{
list(nodes = data.frame, links = data.frame)
}
\description{
Attaches numServers to the given infrastructure switches
}
\note{
the function iterates through all the switches to assign the servers
}
