% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIntensity.R
\name{baseIntensity}
\alias{baseIntensity}
\title{baseIntensity}
\usage{
baseIntensity(regions, regionId, samples)
}
\arguments{
\item{regions}{absolute path to the JSON file with the regions locations
and census, or list with already read JSON file}

\item{regionId}{identifier of the region within the region_json}

\item{samples}{number of longitude and latitude samples to grid the region}
}
\value{
list with the intensity matrix, the axis of longitude/latitude
        samples list(intenMatrix, lonAxis, latAxis)
}
\description{
Generates the template intensity function matrix for the area.
That is, still a factor parameter must be multiplying for specific
generation purposes.
}
