% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coboNodes-data.R
\docType{data}
\name{coboNodes}
\alias{coboNodes}
\title{Cobo Calleja 5G MEC infrastructure nodes data.frame}
\format{An object of class \code{"list"};  TODO - specify better the format}
\source{
\href{https://ieeexplore.ieee.org/document/8667076}{Region 5g MEC case study}
}
\usage{
data(coboNodes)
}
\description{
Cobo Calleja 5G MEC infrastructure nodes data.frame
}
\keyword{datasets}
