% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing.R
\name{stackAndFill}
\alias{stackAndFill}
\title{stackAndFill}
\usage{
stackAndFill(dfA, dfB)
}
\arguments{
\item{dfA}{data.frame A}

\item{dfB}{data.frame B}
}
\value{
stacked data.frame
}
\description{
Stack two dataframes and fill with NULL values not present
}
