% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{write5GtoGraph}
\alias{write5GtoGraph}
\title{write5GtoGraph}
\usage{
write5GtoGraph(m1Assoc, m1Coords, m1AccAssocs, accCentCoords, m2Assocs,
  m2Switches, m2AggAssocs, aggCentCoords, m3Assocs, m3Switches, format,
  file)
}
\arguments{
\item{m1Assoc}{data.frame cell coordinates and the M1 ids}

\item{m1Coords}{data.frame M1 coordinates and their ids}

\item{m1AccAssocs}{data.frame M1 coordinates and the access ring ids}

\item{accCentCoords}{data.frame access ring center coordinates and their ids}

\item{m2Assocs}{data.frame access ring coords and associated M2 ids}

\item{m2Switches}{data.frame M2 switches coordinates and their ids}

\item{m2AggAssocs}{data.frame M2 coordinates and aggregation rings ids}

\item{aggCentCoords}{data.frame aggregation ring centers and their ids}

\item{m3Assocs}{data.frame aggregation ring centers and M3 ids}

\item{m3Switches}{data.frame with M3 coordinates and their ids}

\item{format}{output graph format}

\item{file}{path to the file where the graph is stored}
}
\description{
Writes the 5G scenario as a graph. Each data.frame parameter
contains lat and lon coordinates and group columns to indicate associations
among elements.
}
